+++ DatadogInternal/Tests/Context/DeviceInfoTests.swift
@@ -32,4 +32,46 @@ class DeviceInfoTests: XCTestCase {
         XCTAssertEqual(info.osVersion, randomOSVersion)
         XCTAssertNotNil(info.osBuildNumber)
     }
+
+    func testDeviceType() {
+        // Given
+        let iPhone = UIDeviceMock(model: "iPhone14,5", systemName: "iOS")
+        let iPod = UIDeviceMock(model: "iPod7,1", systemName: "iOS")
+        let iPad = UIDeviceMock(model: "iPad12,1", systemName: "iPadOS")
+        let appleTV1 = UIDeviceMock(model: "J305AP", systemName: "tvOS")
+        let appleTV2 = UIDeviceMock(model: "AppleTV14,1 Simulator", systemName: "tvOS")
+        let other = UIDeviceMock(model: "RealityDevice14,1", systemName: "visionOS")
+
+        // When / Then
+        func when(device: UIDeviceMock) -> DeviceInfo {
+            return DeviceInfo(processInfo: ProcessInfoMock(), device: device)
+        }
+
+        XCTAssertEqual(when(device: iPhone).type, .iPhone)
+        XCTAssertEqual(when(device: iPod).type, .iPod)
+        XCTAssertEqual(when(device: iPad).type, .iPad)
+        XCTAssertEqual(when(device: appleTV1).type, .appleTV)
+        XCTAssertEqual(when(device: appleTV2).type, .appleTV)
+        XCTAssertEqual(when(device: other).type, .other(modelName: "RealityDevice14,1 Simulator", osName: "visionOS"))
+    }
+
+    func testOSVersionMajor() {
+        // When
+        func when(systemVersion: String) -> DeviceInfo {
+            return DeviceInfo(
+                processInfo: ProcessInfoMock(),
+                device: UIDeviceMock(systemVersion: systemVersion)
+            )
+        }
+
+        // Then
+        XCTAssertEqual(when(systemVersion: "15.4.1").osVersion, "15.4.1")
+        XCTAssertEqual(when(systemVersion: "15.4.1").osVersionMajor, "15")
+
+        XCTAssertEqual(when(systemVersion: "17.0").osVersion, "17.0")
+        XCTAssertEqual(when(systemVersion: "17.0").osVersionMajor, "17")
+
+        XCTAssertEqual(when(systemVersion: "18").osVersion, "18")
+        XCTAssertEqual(when(systemVersion: "18").osVersionMajor, "18")
+    }
 }