+++ DatadogRUM/Sources/RUMMonitor/Scopes/RUMResourceScope.swift
@@ -188,7 +188,7 @@ internal class RUMResourceScope: RUMScope {
             date: resourceStartTime.addingTimeInterval(serverTimeOffset).timeIntervalSince1970.toInt64Milliseconds,
             device: .init(context: context, telemetry: dependencies.telemetry),
             display: nil,
-            os: .init(context: context),
+            os: .init(device: context.device),
             resource: .init(
                 connect: resourceMetrics?.connect.map { metric in
                     .init(
@@ -321,7 +321,7 @@ internal class RUMResourceScope: RUMScope {
                 wasTruncated: nil
             ),
             freeze: nil,
-            os: .init(context: context),
+            os: .init(device: context.device),
             service: context.service,
             session: .init(
                 hasReplay: context.hasReplay,