+++ DatadogRUM/Sources/RUMMonitor/Scopes/RUMUserActionScope.swift
@@ -174,7 +174,7 @@ internal class RUMUserActionScope: RUMScope, RUMContextProvider {
             date: actionStartTime.addingTimeInterval(serverTimeOffset).timeIntervalSince1970.toInt64Milliseconds,
             device: .init(context: context, telemetry: dependencies.telemetry),
             display: nil,
-            os: .init(context: context),
+            os: .init(device: context.device),
             service: context.service,
             session: .init(
                 hasReplay: context.hasReplay,